requirements.txt
Django==3.0.1
django-cors-headers==3.4.0
djangorestframework==3.11.0
mysqlclient==1.4.6
